Develop

+++ src/modules/customer/useCases/deleteCustomer/DeleteCustomerUseCase.ts
@@ -1,7 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 
 import { IUCustomersRepository } from '../../repositories/ICustomersRepository';
-import { Customer } from '../../entities/Customer';
 
 @injectable()
 class DeleteCustomerUseCase {
@@ -10,8 +9,20 @@ class DeleteCustomerUseCase {
     private customerRepository: IUCustomersRepository
   ) {}
 
-  async execute(id: string): Promise<Customer> {
-    return await this.customerRepository.delete(id);
+  async execute(id: string): Promise<any> {
+    if (!id) {
+      return { error: 'Id is required!', status: 400 };
+    }
+
+    try {
+      const foundCustomer = await this.customerRepository.findById(id);
+      if (!foundCustomer) {
+        return { error: 'Not Found!', status: 404 };
+      }
+      return await this.customerRepository.delete(id);
+    } catch (error) {
+      return { error: 'Not found', status: 404 };
+    }
   }
 }
